requirements.txt
django
daphne
channels
channels-redis
uvicorn
gunicorn
asgiref
motor
pymongo
dnspython
python-dotenv
